Preliminary code addition. DOES NOT WORK
Preliminary addition of code to git repo. Note that as of this commit, the code is still non-functional. It is almost complete.

Currently Working:
*Read log file
*Parse log file header

Mostly Working:
*Parsing raw data

Not Working
*Applying Coefficents to raw data values
*Formatting compensated data

Using the coefficent data has not been tested, as it requires proper reading of the raw data in order to work.

Potential TODOs:
*Output compoensated data in csv file
*Graph data from sensors?